cursor-marketplace/agent-memory/pyproject.toml
[build-system]
requires = ["uv_build>=0.10.4,<0.11.0"]
build-backend = "uv_build"

[project]
name = "agent-memory"
version = "1.1.0"
description = "Long-term temporal knowledge graph memory for AI agents, with CLI and MCP interfaces"
readme = "README.md"
requires-python = ">=3.12"
license = "Apache-2.0"
keywords = ["mcp", "model-context-protocol", "memory", "knowledge-graph", "graphiti", "neo4j"]
dependencies = [
    "graphiti-core[anthropic,voyageai]>=0.28",
    "fastmcp>=3.0.0,<4",
    "mcp-common>=0.6.0",
    "typer>=0.15.0",
    "rich>=14.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "schedule>=1.2",
    "uvicorn>=0.34",
]

[tool.uv.sources]
mcp-common = { git = "https://github.com/vhspace/mcp-common", rev = "d1d340fee3c2a85b009c917d12b16f8ec8a4803c" }

[project.scripts]
agent-memory = "agent_memory.server:main"
mem = "agent_memory.cli:app"

[dependency-groups]
dev = [
    "mcp-common[testing]",
    "pytest>=8.0",
    "anyio>=4.0",
    "pytest-httpx>=0.34",
    "inline-snapshot>=0.15",
    "mypy>=1.13",
    "ruff>=0.8",
    "pre-commit>=4.0",
]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = [
    "src/agent_memory/__init__.py:__version__",
    ".cursor-plugin/plugin.json:\"version\"",
    ".claude-plugin/plugin.json:\"version\"",
]
branch = "main"
upload_to_vcs_release = true
tag_format = "v{version}"

[tool.semantic_release.commit_parser_options]
allowed_tags = ["feat", "fix", "docs", "chore", "refactor", "test", "ci", "perf"]
minor_tags = ["feat"]
patch_tags = ["fix", "perf"]

[tool.semantic_release.changelog]
changelog_file = "CHANGELOG.md"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E402", "E501", "SIM102", "SIM105", "SIM108", "SIM117", "UP047"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "e2e: end-to-end tests (requires Neo4j + API keys)",
]
